+++ Code/OSUI/OSDropDown.cpp
@@ -49,7 +49,7 @@ void OSDropDown::Init( int32_t x, int32_t y, uint32_t w, uint32_t h )
                                   "ComboBox",
                                   CBS_DROPDOWNLIST | CBS_HASSTRINGS | WS_CHILD | WS_OVERLAPPED | WS_VISIBLE,
                                   x, y,
-                                  w, h,
+                                  (int32_t)w, (int32_t)h,
                                   (HWND)m_Parent->GetHandle(),
                                   nullptr,
                                   nullptr, // TODO: ??? m_HInstance
