+++ Code/Tools/FBuild/FBuildCore/Helpers/FBuildStats.cpp
@@ -130,7 +130,7 @@ void FBuildStats::OutputSummary() const
         for ( size_t i=0; i<itemsToDisplay; ++i )
         {
             const Node * n = m_NodesByTime[ i ];
-            output.AppendFormat( "%-9.3f %s\n", ( (float)n->GetProcessingTime() / 1000.0f ), n->GetName().Get() );
+            output.AppendFormat( "%-9.3f %s\n", (double)( (float)n->GetProcessingTime() / 1000.0f ), n->GetName().Get() );
         }
         output += "\n";
     }
@@ -184,7 +184,7 @@ void FBuildStats::OutputSummary() const
         {
             hitPerc = ( (float)hits / float( hits + misses ) * 100.0f );
         }
-        output.AppendFormat( " - Hits       : %u (%2.1f %%)\n", hits, hitPerc );
+        output.AppendFormat( " - Hits       : %u (%2.1f %%)\n", hits, (double)hitPerc );
         output.AppendFormat( " - Misses     : %u\n", misses );
         output.AppendFormat( " - Stores     : %u\n", stores );
     }
@@ -197,10 +197,10 @@ void FBuildStats::OutputSummary() const
     float totalRemoteCPUInSeconds = (float)( (double)m_TotalRemoteCPUTimeMS / (double)1000 );
     FormatTime( totalLocalCPUInSeconds, buffer );
     float localRatio = ( totalLocalCPUInSeconds / m_TotalBuildTime );
-    output.AppendFormat( " - Local CPU  : %s (%2.1f:1)\n", buffer.Get(), localRatio );
+    output.AppendFormat( " - Local CPU  : %s (%2.1f:1)\n", buffer.Get(), (double)localRatio );
     FormatTime( totalRemoteCPUInSeconds, buffer );
     float remoteRatio = ( totalRemoteCPUInSeconds / m_TotalBuildTime );
-    output.AppendFormat( " - Remote CPU : %s (%2.1f:1)\n", buffer.Get(), remoteRatio );
+    output.AppendFormat( " - Remote CPU : %s (%2.1f:1)\n", buffer.Get(), (double)remoteRatio );
     output += "-----------------------------------------------------------------\n";
 
     OUTPUT( "%s", output.Get() );
@@ -321,7 +321,7 @@ void FBuildStats::FormatTime( float timeInSeconds , AString & buffer ) const
         buffer += temp;
     }
 
-    temp.Format( "%2.3fs", timeInSeconds );
+    temp.Format( "%2.3fs", (double)timeInSeconds );
     buffer += temp;
 }
